// api/models/messageStatus.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * MagicMessenger
 * OpenAPI spec version: v1
 */

export type MessageStatus = (typeof MessageStatus)[keyof typeof MessageStatus];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const MessageStatus = {
  Sent: "Sent",
  Delivered: "Delivered",
  Seen: "Seen",
} as const;