// api/models/messageType.ts
/**
 * Generated by orval v7.14.0 🍺
 * Do not edit manually.
 * MagicMessenger
 * OpenAPI spec version: v1
 */

export type MessageType = (typeof MessageType)[keyof typeof MessageType];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const MessageType = {
  Text: "Text",
  Image: "Image",
  Video: "Video",
  Audio: "Audio",
} as const;